requirements.txt
argon2-cffi==20.1.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.24
Pillow==8.2.0
protobuf==3.15.8
pycparser==2.20
PyPDF2==1.26.0
reportlab==3.5.67
six==1.15.0
Werkzeug==1.0.1
